Make sure we have valid package caches for Debian
Be cause we don't want to manage apt or yum in our role, we need to
make sure each has a valid cache before running the role.

Currently we are only seeing this issue under Ubuntu Trusty, but we
should keep an eye out to see if it affects Centos.

Change-Id: Ib365b89284e18935389a86bb9d3904bfa3c83cc7
Signed-off-by: Paul Belanger <pabelanger@redhat.com>